Initial Linux port (#1)
* Initial Linux port. Had some problems with some of the dependencies; libffi and openssl 3 are expected to be in LIBRARY_PATH now.
* Tidy up for OS X
* Minor cleanup.
* Move con4m_ headers into a con4m subdir.
* Moved libcon4m c sources into their own subdir, and moved vendor headers into a separate dir.
* Simple script for automating meson build targets, that will auto-add the c0 lib dir from .local
* Have meson build static libraries by default.

+++ src/con4m/hex.c
@@ -1,4 +1,4 @@
-#include <con4m_hex.h>
+#include <con4m.h>
 
 const uint8_t hex_map[16] = { '0', '1', '2', '3', '4', '5', '6', '7',
                               '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
@@ -38,7 +38,7 @@ add_offset(char **optr, uint64_t start_offset, uint64_t offset_len,
     char buf[offset_len];
 
 
-    for (int i = 0; i < offset_len; i++) {
+    for (uint64_t i = 0; i < offset_len; i++) {
 	buf[i] = '0';
     }
 
@@ -158,7 +158,7 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
 	num_lines -= 1;
     }
 
-    for (int i = 0; i < num_lines; i++) {
+    for (uint64_t i = 0; i < num_lines; i++) {
 	add_offset(&outptr, start_offset, offset_len, i, chars_per_line);
 
 	// The inner loop is for quads.
@@ -168,14 +168,17 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
             *outptr++ = ' ';
+
             c         = *inptr++;
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
             *outptr++ = ' ';
+
             c         = *inptr++;
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
             *outptr++ = ' ';
+
             c         = *inptr++;
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
@@ -184,7 +187,7 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
 	}
 	// Now for any ASCII-printable stuff, we emit it, or a '.' if not.
 	// lineptr is pointing at the first char we need to show.
-	for (int j = 0; j < chars_per_line; j++) {
+	for (uint64_t j = 0; j < chars_per_line; j++) {
 	    ASCIICHAR();
 	}
 
@@ -203,7 +206,7 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
 	char *stopptr = outptr + (chars_per_line/4)*13;
 
 	// Now, print any full groups of 4.
-	for (int i = 0; i < remainder / 4; i++) {
+	for (uint64_t i = 0; i < remainder / 4; i++) {
             c         = *inptr++;
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
@@ -223,7 +226,7 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
 	    *outptr++ = ' ';
 	}
 	// Now, print any leftover chars.
-	for (int i = 0; i < remainder % 4; i++) {
+	for (uint64_t i = 0; i < remainder % 4; i++) {
             c         = *inptr++;
             *outptr++ = hex_map[(c >> 4)];
             *outptr++ = hex_map[c & 0x0f];
@@ -234,11 +237,11 @@ chexl(void *ptr, unsigned int len, uint64_t start_offset,
 	while (outptr < stopptr) {
 	    *outptr++ = ' ';
 	}
-	for (int i = 0; i < remainder; i++) {
+	for (uint64_t i = 0; i < remainder; i++) {
 	    ASCIICHAR();
 	}
 	// Now, pad the rest of the line w/ spaces;
-	for (int i = remainder; i < chars_per_line; i++) {
+	for (uint64_t i = remainder; i < chars_per_line; i++) {
 	    *outptr++ = ' ';
 	}
 	*outptr = '\n';
